requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework==3.12.4
geojson==2.5.0
pkg-resources==0.0.0
pytz==2021.1
Shapely==1.7.1
sqlparse==0.4.1
